feat:用户管理

+++ server/services/modules/users/users.go
@@ -57,9 +57,13 @@ func (u *userService) FindUserByParams(params *dto.QueryUsersParams) ([]dto.User
 }
 
 func (u *userService) UpdateUser(params *dto.UpdateUserRequest, id string) error {
-	_, exist := u.FindUserById(id)
+	user, exist := u.FindUserById(id)
 	if !exist {
 		return errors.New("资源不存在")
+	}
+	// 超级管理员无法修改
+	if user.IsAdmin == 1 {
+		return errors.New("超级管理员无法修改")
 	}
 	if params.RoleID != nil {
 		// 给用户分配角色信息
@@ -72,9 +76,12 @@ func (u *userService) UpdateUser(params *dto.UpdateUserRequest, id string) error
 }
 
 func (u *userService) DeleteUser(id string) error {
-	_, exist := u.FindUserById(id)
+	user, exist := u.FindUserById(id)
 	if !exist {
 		return errors.New("资源不存在")
+	}
+	if user.IsAdmin == 1 {
+		return errors.New("超级管理员无法删除")
 	}
 	return repositories.UserRepositorysModules.DeleteUser(id)
 }
